Updating to plotly.js v1.45.0
Forked the repo in order to manually update the plotly.js version
myself. This is needed because RStudio Connect uses the packrat package
to manage which versions are installed. Since I can't reference a local
package change, I need to host it from Git (and I can't use the
Enterprise server...)